part 6

+++ src/TileMap/TileMap.java
@@ -151,12 +151,12 @@ public class TileMap {
         return tileSize;
     }
 
-    public int getX(){
-        return (int)x;
+    public double getX(){
+        return x;
     }
 
-    public int getY(){
-        return (int)y;
+    public double getY(){
+        return y;
     }
 
     public int getWidth(){
